requirements.txt
# ──────────────────────────────────────────────
# Python Dependencies
# E-commerce Data Pipeline with Automated ETL
# ──────────────────────────────────────────────
# Install: pip install -r requirements.txt
# ──────────────────────────────────────────────

# Core ETL Libraries
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0

# Database Drivers
psycopg2-binary>=2.9.0       # PostgreSQL connector

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Apache Airflow (Orchestration — install separately if needed)
# apache-airflow>=2.7.0

# Data Formats
pyarrow>=14.0.0              # Parquet file support
openpyxl>=3.1.0              # Excel file support

# Configuration
pyyaml>=6.0.0                # YAML config parser
python-dotenv>=1.0.0         # .env file support

# Testing
pytest>=7.4.0
